+++ gendarme/rules/Gendarme.Rules.Performance/OverrideValueTypeDefaultsRule.cs
@@ -98,10 +98,14 @@ namespace Gendarme.Rules.Performance
     private const string MissingImplementationMessage = "Missing type-specific implementation for '{0}'. {1}";
     private const string MissingOperatorsMessage = "If your langage supports overloading operators then you should implement the equality (==) and inequality (!=) operators.";
 
+    // F# compiler generates types with names like T8_314Bytes@, which are value types that do not override Equals and GetHashCode, but are not user defined.
+    // So we exclude them from the rule.
+    private static readonly System.Text.RegularExpressions.Regex Matcher = new System.Text.RegularExpressions.Regex("^T\\d+_\\d+Bytes@$");
+
     public RuleResult CheckType(TypeDefinition type)
     {
       // rule applies only to ValueType, except enums and generated code
-      if (!type.IsValueType || type.IsEnum || type.IsGeneratedCode())
+      if (!type.IsValueType || type.IsEnum || type.IsGeneratedCode() || Matcher.IsMatch(type.Name))
         return RuleResult.DoesNotApply;
 
       // note: no inheritance check required since we're dealing with structs